# social-flow-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Social Flow Backend API
  description: |
    A comprehensive social media backend API combining YouTube and Twitter features
    with advanced AI/ML capabilities. This API supports video streaming, social posts,
    real-time interactions, content moderation, analytics, and machine learning features.
  version: 1.0.0
  contact:
    name: Social Flow Team
    email: support@socialflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.socialflow.com/api/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
  /auth/register:
    post:
      tags:
        - authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request - validation error
        '409':
          description: Conflict - user already exists
  /auth/login:
    post:
      tags:
        - authentication
      summary: Login user
      description: Authenticate user and return access token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Username or email
                password:
                  type: string
                  format: password
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Unauthorized - invalid credentials
  /auth/refresh:
    post:
      tags:
        - authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
        - display_name
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 100
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        bio:
          type: string
          maxLength: 500
        avatar_url:
          type: string
          format: uri
        website:
          type: string
          format: uri
        location:
          type: string
          maxLength: 100
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
          format: uri
        website:
          type: string
          format: uri
        location:
          type: string
        is_active:
          type: boolean
        is_verified:
          type: boolean
        is_banned:
          type: boolean
        is_suspended:
          type: boolean
        followers_count:
          type: integer
        following_count:
          type: integer
        posts_count:
          type: integer
        videos_count:
          type: integer
        total_views:
          type: integer
        total_likes:
          type: integer
        privacy_level:
          type: string
          enum: [public, friends, private]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
    Token:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
...